requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi @ file:///C:/Windows/Temp/abs_6808y9x40v/croots/recipe/cffi_1659598653989/work
charset-normalizer==2.1.1
colorama==0.4.5
cycler==0.11.0
efficientnet-pytorch==0.7.1
frozenlist==1.3.1
fsspec==2022.8.2
future @ file:///C:/ci/future_1607568713721/work
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
idna==3.3
imageio==2.21.3
imgviz==1.5.1
importlib-metadata==4.12.0
kiwisolver==1.4.4
labelme==5.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.2.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multidict==6.0.2
munch==2.5.0
natsort==8.2.0
networkx==2.8.6
numpy @ file:///C:/Windows/Temp/abs_2a1e1vbeag/croots/recipe/numpy_and_numpy_base_1659432712056/work
numpy-stl==2.17.1
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pretrainedmodels==0.7.4
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.1
pyglet==1.5.27
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-utils==3.3.3
pytorch-lightning==1.5.4
PyWavelets==1.3.0
PyYAML==6.0
QtPy==2.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Rtree==1.0.0
scipy==1.9.1
segmentation-models-pytorch==0.3.0
Shapely==1.8.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
tifffile==2022.8.12
timm==0.4.12
torch==1.12.1
torchmetrics==0.9.3
torchvision==0.13.1
tqdm==4.64.1
trimesh==3.15.1
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3==1.26.12
Werkzeug==2.2.2
wincertstore==0.2
yarl==1.8.1
zipp==3.8.1
